feat(sqlite): batch 91 — strict opt-in, dataSourceExists via pragma_table_list
- Add SqliteOpenConfig.strict on the activesupport driver spec. Documented as a
  no-op for better-sqlite3 (compiled with SQLITE_DQS=0, no sqlite3_db_config
  binding) so future drivers can honor it without churn.
- Align Sqlite3Adapter#dataSourceExists with #tableExists: route the bare-name
  case through pragma_table_list, matching Rails data_source_sql. Schema-
  qualified names still fall back to <schema>.sqlite_master.
- Extract assertLogged into adapters/sqlite3/test-helper.ts so future sqlite3
  test files can share the sql.active_record subscription glue.
